+++ data/time_window_avg_test.go
@@ -7,6 +7,10 @@ import (
 )
 
 func TestNewPoint(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test in short mode")
+	}
+
 	var avgPoint Point
 
 	point := Point{